tweet_coll/requirements.txt
tweepy==4.4.0
python-dotenv
pymongo
dnspython
